requirements.txt
absl-py==1.0.0
alabaster==0.7.12
alembic==1.7.6
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
ast-decompiler==0.4.0
astroid==2.9.3
asttokens==2.0.5
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.10.0
black==22.1.0
bleach==4.1.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cloudpickle==2.0.0
commonmark==0.9.1
coverage==6.3.2
cucim==22.2.0
cycler==0.11.0
databricks-cli==0.16.4
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
docker==5.0.3
docutils==0.18.1
einops==0.4.0
entrypoints==0.4
executing==0.8.2
fastremap==1.12.2
filelock==3.6.0
fill-voids==2.0.2
flake8==4.0.1
flake8-bugbear==22.1.11
flake8-comprehensions==3.8.0
flake8-executable==2.1.1
flake8-polyfill==1.0.2
flake8-pyi==22.2.0
Flask==2.0.3
fonttools==4.29.1
gdown==4.3.1
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.0
google-auth-oauthlib==0.4.6
greenlet==1.1.2
grpcio==1.44.0
gunicorn==20.1.0
huggingface-hub==0.4.0
idna==3.3
ignite==1.1.0
imagecodecs==2022.2.22
imageio==2.16.0
imagesize==1.3.0
importlab==0.7
importlib-metadata==4.11.1
importlib-resources==5.4.0
ipykernel==6.9.1
ipython==8.1.0
ipython-genutils==0.2.0
isort==5.10.1
itk==5.2.1.post1
itk-core==5.2.1.post1
itk-filtering==5.2.1.post1
itk-io==5.2.1.post1
itk-numerics==5.2.1.post1
itk-registration==5.2.1.post1
itk-segmentation==5.2.1.post1
itsdangerous==2.1.0
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyter-server==1.13.5
jupyterlab==3.3.1
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
kiwisolver==1.3.2
lazy-object-proxy==1.7.1
libcst==0.4.1
lmdb==1.3.0
lungmask==0.2.11
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.1.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.7.0
mistune==0.8.4
mlflow==1.23.1
monai==0.8.1
msgpack==1.0.3
mypy==0.931
mypy-extensions==0.4.3
nbclassic==0.3.6
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
networkx==2.6.3
nibabel==3.2.2
ninja==1.10.2.3
notebook==6.4.8
notebook-shim==0.1.0
numpy==1.22.2
oauthlib==3.2.0
opencv-python==4.5.5.64
openslide-python==1.1.2
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parameterized==0.8.1
parso==0.8.3
pathspec==0.9.0
pep8-naming==0.12.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.1
prometheus-client==0.13.1
prometheus-flask-exporter==0.18.7
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydicom==2.2.2
pyflakes==2.4.0
Pygments==2.11.2
pylint==2.12.2
pynrrd==0.4.2
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytorch-ignite==0.4.8
pytype==2022.2.23
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
querystring-parser==1.2.4
ray==1.10.0
recommonmark==0.6.0
redis==4.1.4
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
sacremoses==0.0.47
scikit-image==0.19.2
scipy==1.8.0
Send2Trash==1.8.0
SimpleITK==2.1.1
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
snowballstemmer==2.2.0
soupsieve==2.3.1
Sphinx==3.5.3
sphinx-autodoc-typehints==1.11.1
sphinx-rtd-theme==0.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.31
sqlparse==0.4.2
stack-data==0.2.0
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5
terminado==0.13.3
testpath==0.6.0
tifffile==2022.2.9
tokenizers==0.11.5
toml==0.10.2
tomli==2.0.1
torch==1.10.2
torchvision==0.11.3
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.16.2
types-pkg-resources==0.1.3
typing-extensions==4.0.1
typing-inspect==0.7.1
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.1
Werkzeug==2.0.3
wrapt==1.13.3
zipp==3.7.0
